Fix env variable dependency issue resulting in us having to include AWS ACCOUNT for every lambda
