+++ bin/resports-aws-cdk.ts
@@ -2,7 +2,7 @@
 import "source-map-support/register";
 import * as cdk from "aws-cdk-lib";
 import { ResportsAwsCdkStack } from "../lib/resports-aws-cdk-stack";
-import { env } from "../config/database";
+import { env } from "../config/env";
 
 const app = new cdk.App();
 new ResportsAwsCdkStack(app, "ResportsAwsCdkStack", {
