+++ lib/resports-aws-cdk-stack.ts
@@ -7,7 +7,7 @@ import { UserApiRoutes } from "../constructs/user-api-routes";
 import { SavedChannelApiRoutes } from "../constructs/saved-channels-api-routes";
 import { VPC } from "../constructs/vpc";
 import { PostgresDatabase } from "../constructs/database";
-import { env } from "../config/database";
+import { databaseConfig } from "../config/database";
 import { SubnetType } from "aws-cdk-lib/aws-ec2";
 import { NodejsFunction } from "aws-cdk-lib/aws-lambda-nodejs";
 import { join } from "path";
@@ -42,13 +42,7 @@ export class ResportsAwsCdkStack extends cdk.Stack {
       entry: join(__dirname, "/../lambdas", "initialiseDatabase.ts"),
       vpc: vpc.vpc,
       vpcSubnets: { subnetType: SubnetType.PRIVATE_ISOLATED },
-      environment: {
-        DATABASE_HOST: env.DATABASE_HOST,
-        DATABASE_NAME: env.DATABASE_NAME,
-        DATABASE_PASSWORD: env.DATABASE_PASSWORD,
-        DATABASE_USER: env.DATABASE_USER,
-        AWS_ACCOUNT: env.AWS_ACCOUNT,
-      },
+      environment: databaseConfig,
       timeout: Duration.seconds(30),
       runtime: lambda.Runtime.NODEJS_16_X,
     });
